fixes bug with stale values from previous expressions
only affects users who use consecutive expressions in
which the result of one epxression is used in the next.

```
+  --group-expr "expr1: ..." \
+  --group-expr "expr2:('expr1' in INFO) && ..."\
```

here, in some cases, even if expr1 did not pass, then it would
appear as though it did in the expr2.

This only happened in group and family expressions. trio expressions
had protected against this.